backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.8.1
attrs==25.4.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==45.0.3
dj-database-url==2.1.0
dj-rest-auth==7.0.1
Django==5.2.1
django-admin-rangefilter==0.13.2
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.1
django-rangefilter==0.9.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
multidict==6.7.0
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
paystack==1.5.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.4.0
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
stripe==8.11.0
tabulate==0.9.0
twilio==9.8.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.22.0
weasyprint==66.0
dnspython==2.6.1
# Document parsing for exam uploads
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0
